requirements.txt
python-telegram-bot==13.1
libdw==4.3.0  